perform fft on audio samples with fftw library

+++ include/asfproject/process_audio.h
@@ -76,7 +76,7 @@ private:
   bool loadWaveFile();
 
 public:
-  std::vector<float> samples;
+  std::vector<double> samples;
 
   ProcessAudio();
   explicit ProcessAudio(const std::string &);
